Rework all the tests to make it easier
The following reworks all the tests so the mocks are correctly
made at the right time. This means that we have to move some of
the tests out of the base suite to call them with the right api.
